Split up pattern search test classes
For length
